фиксы меню-вью
